ANDROID: Merge linux-kselftest from upstream tag v6.15.7
Merge squash the linux-kselftest suite from the upstream Linux kernel stable tag v6.15.7.

This large update brings numerous new tests, bug fixes, and infrastructure improvements.

Highlights of this merge include:

- bpf: Major overhaul of the BPF selftests.
- Added new test suites for various features: arena spinlocks

Bug: 422759683
Test: N/A

+++ tools/testing/selftests/timers/set-timer-lat.c
@@ -75,9 +75,9 @@ char *clockstring(int clockid)
 
 long long timespec_sub(struct timespec a, struct timespec b)
 {
-	long long ret = (long long)NSEC_PER_SEC * b.tv_sec + b.tv_nsec;
+	long long ret = NSEC_PER_SEC * b.tv_sec + b.tv_nsec;
 
-	ret -= (long long)NSEC_PER_SEC * a.tv_sec + a.tv_nsec;
+	ret -= NSEC_PER_SEC * a.tv_sec + a.tv_nsec;
 	return ret;
 }
 
@@ -91,7 +91,7 @@ void sigalarm(int signo)
 	alarmcount++;
 
 	delta_ns = timespec_sub(start_time, ts);
-	delta_ns -= (long long)NSEC_PER_SEC * TIMER_SECS * alarmcount;
+	delta_ns -= NSEC_PER_SEC * TIMER_SECS * alarmcount;
 
 	if (delta_ns < 0)
 		timer_fired_early = 1;
